imitation-learning/iq_learn/requirements.txt
# gym[box2d]==0.17.1
# hydra-core==1.0.6
# stable_baselines3==1.0
# tensorboard==2.4.0
# tensorboardX==2.1
# torch<2.0.0
# torchvision==0.10
# tornado==5.1.1
# tqdm==4.42.1
# wandb
# opencv-python
# termcolor
# #atari-py==0.2.6
# #gym_minigrid==1.0.2
# #mujoco_py==2.0.2
nes-py==8.2.1
-e ../../tetris-environments/gym-tetris


absl-py==1.4.0
ale-py==0.8.1
antlr4-python3-runtime==4.8
appdirs==1.4.4
atari-py==0.2.9
attrs==23.1.0
box2d-py==2.3.5
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==1.3.0
cmake==3.26.3
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.34
decorator==4.4.2
docker-pycreds==0.4.0
fasteners==0.18
filelock==3.12.0
fonttools==4.39.3
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
glfw==2.5.9
google-auth==2.17.3
google-auth-oauthlib==1.0.0
grpcio==1.54.0
gym==0.21
gym-minigrid==1.0.2
gym-notices==0.0.8
hydra-core==1.0.6
idna==3.4
imageio==2.28.1
imageio-ffmpeg==0.4.8
importlib-metadata==4.13.0
importlib-resources==5.12.0
iniconfig==2.0.0
Jinja2==3.1.2
kiwisolver==1.4.4
lit==16.0.3
lz4==4.3.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
moviepy==1.0.3
mpmath==1.3.0
mujoco==2.2.0
mujoco-py==2.1.2.14
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
omegaconf==2.0.6
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.1
pathtools==0.1.2
Pillow==9.5.0
pluggy==1.0.0
proglog==0.1.10
protobuf==3.20.3
psutil==5.9.5
py==1.11.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pygame==2.1.0
pyglet==1.5.0
PyOpenGL==3.1.6
pyparsing==3.0.9
pytest==7.0.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
sentry-sdk==1.22.1
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
stable-baselines3==1.8.0
swig==4.1.1
sympy==1.11.1
tensorboard==2.13.0
tensorboard-data-server==0.7.0
tensorboardX==2.6
termcolor==2.3.0
tomli==2.0.1
torch==2.0.0
torchvision==0.15.1
tornado==6.3.1
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wandb==0.15.2
Werkzeug==2.3.3
zipp==3.15.0